% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gs_cite_data.R
\name{get_gs_cite_data}
\alias{get_gs_cite_data}
\title{Get a scholar's Google Scholar data in a usable format.}
\usage{
get_gs_cite_data(id)
}
\arguments{
\item{id}{A Google Scholar ID from a scholar's profile page URL}
}
\value{
A list of data from the scholar's profile, list of publications, and articles citing the scholar's top-cited first-authored publication.
}
\description{
Get a scholar's Google Scholar data in a usable format.
}
\examples{
get_gs_cite_data(id="DpaI7XMAAAAJ")
}
